support for double

+++ GenericJSON/Initialization.swift
@@ -9,10 +9,10 @@ extension JSON {
     /// of those types.
     public init(_ value: Any) throws {
         switch value {
-        case let num as Float:
+        case let num as Double:
             self = .number(num)
         case let num as Int:
-            self = .number(Float(num))
+            self = .number(Double(num))
         case let str as String:
             self = .string(str)
         case let bool as Bool:
@@ -71,7 +71,7 @@ extension JSON: ExpressibleByDictionaryLiteral {
 
 extension JSON: ExpressibleByFloatLiteral {
 
-    public init(floatLiteral value: Float) {
+    public init(floatLiteral value: Double) {
         self = .number(value)
     }
 }
@@ -79,7 +79,7 @@ extension JSON: ExpressibleByFloatLiteral {
 extension JSON: ExpressibleByIntegerLiteral {
 
     public init(integerLiteral value: Int) {
-        self = .number(Float(value))
+        self = .number(Double(value))
     }
 }
